requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.0.6
django-widget-tweaks==1.4.2
factory-boy==2.9.2
Faker==0.8.17
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyBarcode==0.7
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
qrcode==6.0
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
